requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-admin-interface==0.26.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-rest-knox==4.2.0
django-rosetta==0.9.8
djangorestframework==3.14.0
docx==0.2.4
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==9.4.0
polib==1.2.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-docx==0.8.11
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
text-unidecode==1.3
twilio==8.2.2
typing_extensions==4.5.0
urllib3==1.26.15
xlwt==1.3.0
yarl==1.9.2
